// src/interfaces/bank_card_assignment_info.interface.ts
export default interface BankCardAssignmentInfoInterface {
  unique_entity_id: string;
  entity_id: string;
  staff_id: string;
  pan: string;
  expiry_date: string;
  assigned_date: string;
  ik_number: string;
  bg_card_number: string;
  app_version: string;
  imei: string;
  status: string;
  leader_presence_flag: string;
  item_info: string;
  item_id: string;
  card_image: string;  
  card_holder_image: string;
  bank_card_id: string;
  replacement_reason : string;
  portal_card_status: string;
}